Crosshair script works
changed a few settings with the scene such as setting the canvas to scale with screen size and fixed the crosshair so that it actually shows it. Took off camera movement script from camera to test if the crosshair showed up but it can be placed there again if need be.